Update and rename main.py to NQueen.py
Should work as expected
